# Conflicts:
#	backend/tests/Squidex.Domain.Apps.Core.Tests/Squidex.Domain.Apps.Core.Tests.csproj
#	backend/tests/Squidex.Domain.Apps.Entities.Tests/Contents/GraphQL/TestContent.cs
#	backend/tests/Squidex.Domain.Apps.Entities.Tests/Contents/GraphQL/TestSchemas.cs

+++ backend/src/Squidex.Domain.Apps.Core.Model/Assets/AssetMetadata.cs
@@ -37,28 +37,28 @@ public sealed class AssetMetadata : Dictionary<string, JsonValue>
 
     public AssetMetadata SetPixelWidth(int value)
     {
-        this[PixelWidth] = (double)value;
+        this[PixelWidth] = value;
 
         return this;
     }
 
     public AssetMetadata SetPixelHeight(int value)
     {
-        this[PixelHeight] = (double)value;
+        this[PixelHeight] = value;
 
         return this;
     }
 
     public AssetMetadata SetVideoWidth(int value)
     {
-        this[VideoWidth] = (double)value;
+        this[VideoWidth] = value;
 
         return this;
     }
 
     public AssetMetadata SetVideoHeight(int value)
     {
-        this[VideoHeight] = (double)value;
+        this[VideoHeight] = value;
 
         return this;
     }
